+++ examples/tcp-client-sync.rs
@@ -3,8 +3,8 @@ use tokio_modbus::*;
 
 #[cfg(all(feature = "tcp", feature = "sync"))]
 pub fn main() {
-    let addr = "192.168.0.222:502".parse().unwrap();
-    let mut client = SyncClient::connect_tcp(&addr).unwrap();
+    let socket_addr = "192.168.0.222:502".parse().unwrap();
+    let mut client = SyncClient::connect_tcp(&socket_addr).unwrap();
     let buff = client.read_input_registers(0x1000, 7).unwrap();
     println!("Response is '{:?}'", buff);
 }
