+++ examples/tcp-client.rs
@@ -10,9 +10,9 @@ pub fn main() {
 
     let mut core = Core::new().unwrap();
     let handle = core.handle();
-    let addr = "192.168.0.222:502".parse().unwrap();
+    let socket_addr = "192.168.0.222:502".parse().unwrap();
 
-    let task = Client::connect_tcp(&addr, &handle).and_then(|client| {
+    let task = Client::connect_tcp(&socket_addr, &handle).and_then(|client| {
         println!("Fetching the coupler ID");
         client
             .read_input_registers(0x1000, 7)
