+++ src/client/mod.rs
@@ -1,55 +1,87 @@
+#[cfg(feature = "rtu")]
+pub mod rtu;
+
+#[cfg(feature = "tcp")]
+pub mod tcp;
+
 use crate::frame::*;
+
 use futures::prelude::*;
-use crate::service;
-use std::io::{Error, ErrorKind, Result};
-use std::net::SocketAddr;
-use tokio_core::reactor::{Core, Handle};
+use std::io::{Error, ErrorKind};
+use tokio_core::reactor::Handle;
+use tokio_service::Service;
+
+#[cfg(feature = "rtu")]
+use std::io::Result;
+
 #[cfg(feature = "rtu")]
 use tokio_serial::{Serial, SerialPortSettings};
-use tokio_service::Service;
+
+#[cfg(feature = "rtu")]
+use tokio_core::reactor::Core;
+
+#[cfg(feature = "tcp")]
+use std::net::SocketAddr;
 
 /// A transport independent asynchronous client trait.
 pub trait ModbusClient {
-    fn read_coils(&self, _: Address, _: Quantity) -> Box<Future<Item = Vec<Coil>, Error = Error>>;
+    fn call(&self, req: Request) -> Box<dyn Future<Item = Response, Error = Error>>;
+
+    fn read_coils(
+        &self,
+        _: Address,
+        _: Quantity,
+    ) -> Box<dyn Future<Item = Vec<Coil>, Error = Error>>;
     fn read_discrete_inputs(
         &self,
         _: Address,
         _: Quantity,
-    ) -> Box<Future<Item = Vec<Coil>, Error = Error>>;
-    fn write_single_coil(&self, _: Address, _: Coil) -> Box<Future<Item = (), Error = Error>>;
-    fn write_multiple_coils(&self, _: Address, _: &[Coil]) -> Box<Future<Item = (), Error = Error>>;
+    ) -> Box<dyn Future<Item = Vec<Coil>, Error = Error>>;
     fn read_input_registers(
         &self,
         _: Address,
         _: Quantity,
-    ) -> Box<Future<Item = Vec<Word>, Error = Error>>;
+    ) -> Box<dyn Future<Item = Vec<Word>, Error = Error>>;
     fn read_holding_registers(
         &self,
         _: Address,
         _: Quantity,
-    ) -> Box<Future<Item = Vec<Word>, Error = Error>>;
-    fn write_single_register(&self, _: Address, _: Word) -> Box<Future<Item = (), Error = Error>>;
-    fn write_multiple_registers(&self, _: Address, _: &[Word]) -> Box<Future<Item = (), Error = Error>>;
+    ) -> Box<dyn Future<Item = Vec<Word>, Error = Error>>;
     fn read_write_multiple_registers(
         &self,
         _: Address,
         _: Quantity,
         _: Address,
         _: &[Word],
-    ) -> Box<Future<Item = Vec<Word>, Error = Error>>;
+    ) -> Box<dyn Future<Item = Vec<Word>, Error = Error>>;
+
+    fn write_single_coil(&self, _: Address, _: Coil) -> Box<dyn Future<Item = (), Error = Error>>;
+    fn write_multiple_coils(
+        &self,
+        _: Address,
+        _: &[Coil],
+    ) -> Box<dyn Future<Item = (), Error = Error>>;
+    fn write_single_register(
+        &self,
+        _: Address,
+        _: Word,
+    ) -> Box<dyn Future<Item = (), Error = Error>>;
+    fn write_multiple_registers(
+        &self,
+        _: Address,
+        _: &[Word],
+    ) -> Box<dyn Future<Item = (), Error = Error>>;
 }
 
 /// A transport independent synchronous client trait.
 #[cfg(feature = "sync")]
 pub trait SyncModbusClient {
+    fn call(&mut self, req: Request) -> Result<Response>;
+
     fn read_coils(&mut self, _: Address, _: Quantity) -> Result<Vec<Coil>>;
     fn read_discrete_inputs(&mut self, _: Address, _: Quantity) -> Result<Vec<Coil>>;
-    fn write_single_coil(&mut self, _: Address, _: Coil) -> Result<()>;
-    fn write_multiple_coils(&mut self, _: Address, _: &[Coil]) -> Result<()>;
     fn read_input_registers(&mut self, _: Address, _: Quantity) -> Result<Vec<Word>>;
     fn read_holding_registers(&mut self, _: Address, _: Quantity) -> Result<Vec<Word>>;
-    fn write_single_register(&mut self, _: Address, _: Word) -> Result<()>;
-    fn write_multiple_registers(&mut self, _: Address, _: &[Word]) -> Result<()>;
     fn read_write_multiple_registers(
         &mut self,
         _: Address,
@@ -57,16 +89,21 @@ pub trait SyncModbusClient {
         _: Address,
         _: &[Word],
     ) -> Result<Vec<Word>>;
+
+    fn write_single_coil(&mut self, _: Address, _: Coil) -> Result<()>;
+    fn write_multiple_coils(&mut self, _: Address, _: &[Coil]) -> Result<()>;
+    fn write_single_register(&mut self, _: Address, _: Word) -> Result<()>;
+    fn write_multiple_registers(&mut self, _: Address, _: &[Word]) -> Result<()>;
 }
 
 /// A async modbus client implementation.
 pub struct Client {
-    transport: Box<
+    service: Box<
         Service<
             Request = Request,
             Response = Response,
             Error = Error,
-            Future = Box<Future<Item = Response, Error = Error>>,
+            Future = Box<dyn Future<Item = Response, Error = Error>>,
         >,
     >,
 }
@@ -74,46 +111,37 @@ pub struct Client {
 /// A sync modbus client implementation.
 #[cfg(feature = "sync")]
 pub struct SyncClient {
-    client: Client,
+    async_client: Client,
     core: Core,
 }
 
 impl Client {
     #[cfg(feature = "tcp")]
     pub fn connect_tcp(
-        addr: &SocketAddr,
+        socket_addr: &SocketAddr,
         handle: &Handle,
-    ) -> Box<Future<Item = Client, Error = Error>> {
-        let t = service::tcp::Client::connect(addr, handle).map(|c| Client {
-            transport: Box::new(c),
-        });
-        Box::new(t)
+    ) -> Box<dyn Future<Item = Self, Error = Error>> {
+        Box::new(self::tcp::connect(socket_addr, handle))
     }
+
     #[cfg(feature = "rtu")]
     pub fn connect_rtu(
         serial: Serial,
         address: u8,
         handle: &Handle,
-    ) -> Box<Future<Item = Client, Error = Error>> {
-        let t = service::rtu::Client::connect(serial, address, handle).map(|c| Client {
-            transport: Box::new(c),
-        });
-        Box::new(t)
-    }
-
-    pub fn call(&self, req: Request) -> Box<Future<Item = Response, Error = Error>> {
-        self.transport.call(req)
+    ) -> Box<dyn Future<Item = Client, Error = Error>> {
+        Box::new(self::rtu::connect(serial, address, handle))
     }
 }
 
 #[cfg(feature = "sync")]
 impl SyncClient {
     #[cfg(feature = "tcp")]
-    pub fn connect_tcp(addr: &SocketAddr) -> Result<SyncClient> {
+    pub fn connect_tcp(socket_addr: &SocketAddr) -> Result<SyncClient> {
         let mut core = Core::new()?;
         let handle = core.handle();
-        let client = core.run(Client::connect_tcp(addr, &handle))?;
-        Ok(SyncClient { client, core })
+        let async_client = core.run(Client::connect_tcp(socket_addr, &handle))?;
+        Ok(SyncClient { async_client, core })
     }
     #[cfg(feature = "rtu")]
     pub fn connect_rtu(
@@ -124,21 +152,21 @@ impl SyncClient {
         let mut core = Core::new()?;
         let handle = core.handle();
         let serial = Serial::from_path_with_handle(tty_path, settings, &handle.new_tokio_handle())?;
-        let client = core.run(Client::connect_rtu(serial, address, &handle))?;
-        Ok(SyncClient { client, core })
-    }
-
-    pub fn call(&mut self, req: Request) -> Result<Response> {
-        self.core.run(self.client.call(req))
+        let async_client = core.run(Client::connect_rtu(serial, address, &handle))?;
+        Ok(SyncClient { async_client, core })
     }
 }
 
 impl ModbusClient for Client {
+    fn call(&self, req: Request) -> Box<dyn Future<Item = Response, Error = Error>> {
+        self.service.call(req)
+    }
+
     fn read_coils(
         &self,
         addr: Address,
         cnt: Quantity,
-    ) -> Box<Future<Item = Vec<Coil>, Error = Error>> {
+    ) -> Box<dyn Future<Item = Vec<Coil>, Error = Error>> {
         Box::new(self.call(Request::ReadCoils(addr, cnt)).and_then(|res| {
             if let Response::ReadCoils(coils) = res {
                 Ok(coils)
@@ -152,7 +180,7 @@ impl ModbusClient for Client {
         &self,
         addr: Address,
         cnt: Quantity,
-    ) -> Box<Future<Item = Vec<Coil>, Error = Error>> {
+    ) -> Box<dyn Future<Item = Vec<Coil>, Error = Error>> {
         Box::new(
             self.call(Request::ReadDiscreteInputs(addr, cnt))
                 .and_then(|res| {
@@ -169,7 +197,7 @@ impl ModbusClient for Client {
         &self,
         addr: Address,
         coil: Coil,
-    ) -> Box<Future<Item = (), Error = Error>> {
+    ) -> Box<dyn Future<Item = (), Error = Error>> {
         Box::new(
             self.call(Request::WriteSingleCoil(addr, coil))
                 .and_then(move |res| {
@@ -189,7 +217,7 @@ impl ModbusClient for Client {
         &self,
         addr: Address,
         coils: &[Coil],
-    ) -> Box<Future<Item = (), Error = Error>> {
+    ) -> Box<dyn Future<Item = (), Error = Error>> {
         let cnt = coils.len();
         Box::new(
             self.call(Request::WriteMultipleCoils(addr, coils.to_vec()))
@@ -210,7 +238,7 @@ impl ModbusClient for Client {
         &self,
         addr: Address,
         cnt: Quantity,
-    ) -> Box<Future<Item = Vec<Word>, Error = Error>> {
+    ) -> Box<dyn Future<Item = Vec<Word>, Error = Error>> {
         Box::new(
             self.call(Request::ReadInputRegisters(addr, cnt))
                 .and_then(move |res| {
@@ -230,7 +258,7 @@ impl ModbusClient for Client {
         &self,
         addr: Address,
         cnt: Quantity,
-    ) -> Box<Future<Item = Vec<Word>, Error = Error>> {
+    ) -> Box<dyn Future<Item = Vec<Word>, Error = Error>> {
         Box::new(
             self.call(Request::ReadHoldingRegisters(addr, cnt))
                 .and_then(move |res| {
@@ -250,7 +278,7 @@ impl ModbusClient for Client {
         &self,
         addr: Address,
         data: Word,
-    ) -> Box<Future<Item = (), Error = Error>> {
+    ) -> Box<dyn Future<Item = (), Error = Error>> {
         Box::new(
             self.call(Request::WriteSingleRegister(addr, data))
                 .and_then(move |res| {
@@ -270,7 +298,7 @@ impl ModbusClient for Client {
         &self,
         addr: Address,
         data: &[Word],
-    ) -> Box<Future<Item = (), Error = Error>> {
+    ) -> Box<dyn Future<Item = (), Error = Error>> {
         let cnt = data.len();
         Box::new(
             self.call(Request::WriteMultipleRegisters(addr, data.to_vec()))
@@ -293,15 +321,14 @@ impl ModbusClient for Client {
         read_cnt: Quantity,
         write_addr: Address,
         write_data: &[Word],
-    ) -> Box<Future<Item = Vec<Word>, Error = Error>> {
+    ) -> Box<dyn Future<Item = Vec<Word>, Error = Error>> {
         Box::new(
             self.call(Request::ReadWriteMultipleRegisters(
                 read_addr,
                 read_cnt,
                 write_addr,
                 write_data.to_vec(),
-            ))
-            .and_then(move |res| {
+            )).and_then(move |res| {
                 if let Response::ReadWriteMultipleRegisters(res) = res {
                     if res.len() != read_cnt as usize {
                         return Err(Error::new(ErrorKind::InvalidData, "invalid response"));
@@ -317,30 +344,40 @@ impl ModbusClient for Client {
 
 #[cfg(feature = "sync")]
 impl SyncModbusClient for SyncClient {
+    fn call(&mut self, req: Request) -> Result<Response> {
+        self.core.run(self.async_client.call(req))
+    }
+
     fn read_coils(&mut self, addr: Address, cnt: Quantity) -> Result<Vec<Coil>> {
-        self.core.run(self.client.read_coils(addr, cnt))
+        self.core.run(self.async_client.read_coils(addr, cnt))
     }
     fn read_discrete_inputs(&mut self, addr: Address, cnt: Quantity) -> Result<Vec<Coil>> {
-        self.core.run(self.client.read_discrete_inputs(addr, cnt))
+        self.core
+            .run(self.async_client.read_discrete_inputs(addr, cnt))
     }
     fn write_single_coil(&mut self, addr: Address, coil: Coil) -> Result<()> {
-        self.core.run(self.client.write_single_coil(addr, coil))
+        self.core
+            .run(self.async_client.write_single_coil(addr, coil))
     }
     fn write_multiple_coils(&mut self, addr: Address, coils: &[Coil]) -> Result<()> {
-        self.core.run(self.client.write_multiple_coils(addr, coils))
+        self.core
+            .run(self.async_client.write_multiple_coils(addr, coils))
     }
     fn read_input_registers(&mut self, addr: Address, cnt: Quantity) -> Result<Vec<Word>> {
-        self.core.run(self.client.read_input_registers(addr, cnt))
+        self.core
+            .run(self.async_client.read_input_registers(addr, cnt))
     }
     fn read_holding_registers(&mut self, addr: Address, cnt: Quantity) -> Result<Vec<Word>> {
-        self.core.run(self.client.read_holding_registers(addr, cnt))
+        self.core
+            .run(self.async_client.read_holding_registers(addr, cnt))
     }
     fn write_single_register(&mut self, addr: Address, data: Word) -> Result<()> {
-        self.core.run(self.client.write_single_register(addr, data))
+        self.core
+            .run(self.async_client.write_single_register(addr, data))
     }
     fn write_multiple_registers(&mut self, addr: Address, data: &[Word]) -> Result<()> {
         self.core
-            .run(self.client.write_multiple_registers(addr, data))
+            .run(self.async_client.write_multiple_registers(addr, data))
     }
     fn read_write_multiple_registers(
         &mut self,
@@ -350,7 +387,7 @@ impl SyncModbusClient for SyncClient {
         write_data: &[Word],
     ) -> Result<Vec<Word>> {
         self.core.run(
-            self.client
+            self.async_client
                 .read_write_multiple_registers(read_addr, read_cnt, write_addr, write_data),
         )
     }
