+++ src/codec/rtu.rs
@@ -1,108 +1,118 @@
-use super::common::*;
-use byteorder::ReadBytesExt;
+use super::*;
+
+use crate::frame::rtu::*;
+
 use bytes::{BigEndian, BufMut, Bytes, BytesMut};
-use crate::frame::*;
 use std::io::{Cursor, Error, ErrorKind, Result};
 use tokio_codec::{Decoder, Encoder};
 
 #[derive(Debug, PartialEq)]
-pub(crate) struct Codec {
-    pub(crate) decoder: RtuDecoder,
+pub(crate) struct RequestDecoder;
+
+#[derive(Debug, PartialEq)]
+pub(crate) struct ResponseDecoder;
+
+#[derive(Debug, PartialEq)]
+pub(crate) struct ClientCodec {
+    pub(crate) decoder: ResponseDecoder,
+}
+
+impl Default for ClientCodec {
+    fn default() -> Self {
+        Self {
+            decoder: ResponseDecoder,
+        }
+    }
 }
 
 #[derive(Debug, PartialEq)]
-pub(crate) struct RtuDecoder {
-    codec_type: CodecType,
+pub(crate) struct ServerCodec {
+    pub(crate) decoder: RequestDecoder,
 }
 
-const MIN_ADU_LEN: usize = 1 + 1 + 2; // addr + function + crc
-
-impl Codec {
-    pub fn client() -> Codec {
-        Codec {
-            decoder: RtuDecoder {
-                codec_type: CodecType::Client,
-            },
-        }
-    }
-    pub fn server() -> Codec {
-        Codec {
-            decoder: RtuDecoder {
-                codec_type: CodecType::Server,
-            },
+impl Default for ServerCodec {
+    fn default() -> Self {
+        Self {
+            decoder: RequestDecoder,
         }
     }
 }
 
-fn get_request_payload_len(buf: &BytesMut) -> Result<Option<usize>> {
-    if buf.len() < 2 {
-        // incomplete frame
+fn get_request_pdu_len(adu_buf: &BytesMut) -> Result<Option<usize>> {
+    if adu_buf.len() < 2 {
         return Ok(None);
     }
-    let len = match buf[1] {
-        0x01...0x06 => Some(4),
-        0x07 | 0x0B | 0x0C | 0x11 => Some(0),
+    let fn_code = adu_buf[1];
+    let len = match fn_code {
+        0x01...0x06 => Some(5),
+        0x07 | 0x0B | 0x0C | 0x11 => Some(1),
         0x0F | 0x10 => {
-            if buf.len() > 4 {
-                Some(5 + buf[4] as usize)
+            if adu_buf.len() > 4 {
+                Some(6 + adu_buf[4] as usize)
             } else {
                 // incomplete frame
                 None
             }
         }
-        0x16 => Some(6),
-        0x18 => Some(2),
+        0x16 => Some(7),
+        0x18 => Some(3),
         0x17 => {
-            if buf.len() > 10 {
-                Some(9 + buf[10] as usize)
+            if adu_buf.len() > 10 {
+                Some(10 + adu_buf[10] as usize)
             } else {
                 // incomplete frame
                 None
             }
         }
         _ => {
-            return Err(Error::new(ErrorKind::InvalidData, "invalid data length"));
+            return Err(Error::new(
+                ErrorKind::InvalidData,
+                format!("Invalid function code: 0x{:0>2X}", fn_code),
+            ))
         }
     };
     Ok(len)
 }
 
-fn get_response_payload_len(buf: &BytesMut) -> Result<Option<usize>> {
-    if buf.len() < 2 {
-        // incomplete frame
+fn get_response_pdu_len(adu_buf: &BytesMut) -> Result<Option<usize>> {
+    if adu_buf.len() < 2 {
         return Ok(None);
     }
-    let len = match buf[1] {
+    let fn_code = adu_buf[1];
+    let len = match fn_code {
         0x01...0x04 | 0x0C | 0x17 => {
-            if buf.len() > 2 {
-                Some(1 + buf[2] as usize)
+            if adu_buf.len() > 2 {
+                Some(2 + adu_buf[2] as usize)
             } else {
                 // incomplete frame
                 None
             }
         }
-        0x05 | 0x06 | 0x0B | 0x0F | 0x10 => Some(4),
-        0x07 => Some(1),
-        0x16 => Some(6),
+        0x05 | 0x06 | 0x0B | 0x0F | 0x10 => Some(5),
+        0x07 => Some(2),
+        0x16 => Some(7),
         0x18 => {
-            if buf.len() > 3 {
-                Some(2 + Cursor::new(&buf[2..=3]).read_u16::<BigEndian>()? as usize)
+            if adu_buf.len() > 3 {
+                Some(3 + Cursor::new(&adu_buf[2..=3]).read_u16::<BigEndian>()? as usize)
             } else {
                 // incomplete frame
                 None
             }
         }
-        0x81...0xAB => Some(1),
+        0x81...0xAB => Some(2),
         _ => {
-            return Err(Error::new(ErrorKind::InvalidData, "invalid data length"));
+            return Err(Error::new(
+                ErrorKind::InvalidData,
+                format!("Invalid function code: 0x{:0>2X}", fn_code),
+            ))
         }
     };
     Ok(len)
 }
 
-fn calc_crc(buf: &[u8]) -> u16 {
+fn calc_crc(data: &[u8]) -> u16 {
     let mut crc = 0xFFFF;
-    for x in buf {
+    for x in data {
         crc ^= u16::from(*x);
         for _ in 0..8 {
             if (crc & 0x0001) != 0 {
@@ -116,83 +126,121 @@ fn calc_crc(buf: &[u8]) -> u16 {
     (crc << 8 | crc >> 8)
 }
 
+fn check_crc(adu_data: &[u8], expected_crc: u16) -> Result<()> {
+    let actual_crc = calc_crc(&adu_data);
+    if expected_crc != actual_crc {
+        return Err(Error::new(
+            ErrorKind::InvalidData,
+            format!(
+                "Invalid CRC: expected = 0x{:0>4X}, actual = 0x{:0>4X}",
+                expected_crc, actual_crc
+            ),
+        ));
+    }
+    Ok(())
+}
+
 type ServerAddress = u8;
 
-impl Decoder for RtuDecoder {
+impl Decoder for RequestDecoder {
     type Item = (ServerAddress, Bytes);
     type Error = Error;
 
     fn decode(&mut self, buf: &mut BytesMut) -> Result<Option<(ServerAddress, Bytes)>> {
-        let payload_len: Option<usize> = match self.codec_type {
-            CodecType::Client => get_response_payload_len(buf)?,
-            CodecType::Server => get_request_payload_len(buf)?,
-        }.filter(|payload_len| buf.len() >= MIN_ADU_LEN + payload_len);
-
-        if let Some(payload_len) = payload_len {
-            let mut adu = buf.split_to(payload_len + 2);
-            let crc = buf.split_to(2);
-            let crc = Cursor::new(&crc).read_u16::<BigEndian>()?;
-
-            let expected_crc = calc_crc(&adu);
-            if expected_crc != crc {
-                return Err(Error::new(
-                    ErrorKind::InvalidData,
-                    format!("CRC is not correct: {} instead of {}", crc, expected_crc),
-                ));
+        if let Some(pdu_len) = get_request_pdu_len(buf)? {
+            let adu_len = 1 + pdu_len;
+            if buf.len() >= adu_len + 2 {
+                let mut adu_data = buf.split_to(adu_len);
+                // Read trailing CRC and verify ADU
+                let crc = Cursor::new(&buf.split_to(2)).read_u16::<BigEndian>()?;
+                check_crc(&adu_data, crc)?;
+                let address = adu_data.split_to(1)[0];
+                let pdu_data = adu_data.freeze();
+                return Ok(Some((address, pdu_data)));
             }
-            let address = adu.split_to(1)[0];
-            let data = adu.freeze();
-            Ok(Some((address, data)))
+        }
+        // incomplete frame
+        Ok(None)
+    }
+}
+
+impl Decoder for ResponseDecoder {
+    type Item = (ServerAddress, Bytes);
+    type Error = Error;
+
+    fn decode(&mut self, buf: &mut BytesMut) -> Result<Option<(ServerAddress, Bytes)>> {
+        if let Some(pdu_len) = get_response_pdu_len(buf)? {
+            let adu_len = 1 + pdu_len;
+            if buf.len() >= adu_len + 2 {
+                let mut adu_data = buf.split_to(adu_len);
+                // Read trailing CRC and verify ADU
+                let crc = Cursor::new(&buf.split_to(2)).read_u16::<BigEndian>()?;
+                check_crc(&adu_data, crc)?;
+                let address = adu_data.split_to(1)[0];
+                let pdu_data = adu_data.freeze();
+                return Ok(Some((address, pdu_data)));
+            }
+        }
+        Ok(None)
+    }
+}
+
+impl Decoder for ClientCodec {
+    type Item = ResponseAdu;
+    type Error = Error;
+
+    fn decode(&mut self, buf: &mut BytesMut) -> Result<Option<ResponseAdu>> {
+        if let Some((address, pdu_data)) = self.decoder.decode(buf)? {
+            let hdr = Header { address };
+            let pdu = ResponsePdu::try_from(pdu_data)?;
+            Ok(Some(ResponseAdu { hdr, pdu }))
         } else {
-            // incomplete frame
             Ok(None)
         }
     }
 }
 
-impl Decoder for Codec {
-    type Item = RtuAdu;
+impl Decoder for ServerCodec {
+    type Item = RequestAdu;
     type Error = Error;
 
-    fn decode(&mut self, buf: &mut BytesMut) -> Result<Option<RtuAdu>> {
-        if let Some((address, data)) = self.decoder.decode(buf)? {
-            let pdu = match self.decoder.codec_type {
-                CodecType::Client => {
-                    let res = if data[0] > 0x80 {
-                        Err(ExceptionResponse::try_from(data)?)
-                    } else {
-                        Ok(Response::try_from(data)?)
-                    };
-                    Pdu::Result(res)
-                }
-                CodecType::Server => {
-                    if data[0] > 0x80 {
-                        return Err(Error::new(
-                            ErrorKind::Other,
-                            "A request must not a exception response",
-                        ));
-                    }
-                    let req = Request::try_from(data)?;
-                    Pdu::Request(req)
-                }
-            };
-            Ok(Some(RtuAdu { address, pdu }))
+    fn decode(&mut self, buf: &mut BytesMut) -> Result<Option<RequestAdu>> {
+        if let Some((address, pdu_data)) = self.decoder.decode(buf)? {
+            let hdr = Header { address };
+            let pdu = RequestPdu::try_from(pdu_data)?;
+            Ok(Some(RequestAdu { hdr, pdu }))
         } else {
             Ok(None)
         }
     }
 }
 
-impl Encoder for Codec {
-    type Item = RtuAdu;
+impl Encoder for ClientCodec {
+    type Item = RequestAdu;
     type Error = Error;
 
-    fn encode(&mut self, adu: RtuAdu, buf: &mut BytesMut) -> Result<()> {
-        let RtuAdu { address, pdu } = adu;
-        let pdu: Bytes = pdu.into();
-        buf.reserve(pdu.len() + 3);
-        buf.put_u8(address);
-        buf.put_slice(&*pdu);
+    fn encode(&mut self, adu: RequestAdu, buf: &mut BytesMut) -> Result<()> {
+        let RequestAdu { hdr, pdu } = adu;
+        let pdu_data: Bytes = pdu.into();
+        buf.reserve(pdu_data.len() + 3);
+        buf.put_u8(hdr.address);
+        buf.put_slice(&*pdu_data);
+        let crc = calc_crc(buf);
+        buf.put_u16_be(crc);
+        Ok(())
+    }
+}
+
+impl Encoder for ServerCodec {
+    type Item = ResponseAdu;
+    type Error = Error;
+
+    fn encode(&mut self, adu: ResponseAdu, buf: &mut BytesMut) -> Result<()> {
+        let ResponseAdu { hdr, pdu } = adu;
+        let pdu_data: Bytes = pdu.into();
+        buf.reserve(pdu_data.len() + 3);
+        buf.put_u8(hdr.address);
+        buf.put_slice(&*pdu_data);
         let crc = calc_crc(buf);
         buf.put_u16_be(crc);
         Ok(())
@@ -214,114 +262,113 @@ mod tests {
     }
 
     #[test]
-    fn test_get_request_payload_len() {
+    fn test_get_request_pdu_len() {
         let mut buf = BytesMut::new();
 
         buf.extend_from_slice(&[0x66, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0]);
-        assert!(get_request_payload_len(&buf).is_err());
+        assert!(get_request_pdu_len(&buf).is_err());
 
         buf[1] = 0x01;
-        assert_eq!(get_request_payload_len(&buf).unwrap(), Some(4));
+        assert_eq!(get_request_pdu_len(&buf).unwrap(), Some(5));
 
         buf[1] = 0x02;
-        assert_eq!(get_request_payload_len(&buf).unwrap(), Some(4));
+        assert_eq!(get_request_pdu_len(&buf).unwrap(), Some(5));
 
         buf[1] = 0x03;
-        assert_eq!(get_request_payload_len(&buf).unwrap(), Some(4));
+        assert_eq!(get_request_pdu_len(&buf).unwrap(), Some(5));
 
         buf[1] = 0x04;
-        assert_eq!(get_request_payload_len(&buf).unwrap(), Some(4));
+        assert_eq!(get_request_pdu_len(&buf).unwrap(), Some(5));
 
         buf[1] = 0x05;
-        assert_eq!(get_request_payload_len(&buf).unwrap(), Some(4));
+        assert_eq!(get_request_pdu_len(&buf).unwrap(), Some(5));
 
         buf[1] = 0x06;
-        assert_eq!(get_request_payload_len(&buf).unwrap(), Some(4));
+        assert_eq!(get_request_pdu_len(&buf).unwrap(), Some(5));
 
         buf[1] = 0x07;
-        assert_eq!(get_request_payload_len(&buf).unwrap(), Some(0));
+        assert_eq!(get_request_pdu_len(&buf).unwrap(), Some(1));
 
         // TODO: 0x08
 
         buf[1] = 0x0B;
-        assert_eq!(get_request_payload_len(&buf).unwrap(), Some(0));
+        assert_eq!(get_request_pdu_len(&buf).unwrap(), Some(1));
 
         buf[1] = 0x0C;
-        assert_eq!(get_request_payload_len(&buf).unwrap(), Some(0));
+        assert_eq!(get_request_pdu_len(&buf).unwrap(), Some(1));
 
         buf[1] = 0x0F;
         buf[4] = 99;
-        assert_eq!(get_request_payload_len(&buf).unwrap(), Some(104));
+        assert_eq!(get_request_pdu_len(&buf).unwrap(), Some(105));
 
         buf[1] = 0x10;
         buf[4] = 99;
-        assert_eq!(get_request_payload_len(&buf).unwrap(), Some(104));
+        assert_eq!(get_request_pdu_len(&buf).unwrap(), Some(105));
 
         buf[1] = 0x11;
-        assert_eq!(get_request_payload_len(&buf).unwrap(), Some(0));
+        assert_eq!(get_request_pdu_len(&buf).unwrap(), Some(1));
 
         // TODO: 0x14
 
         // TODO: 0x15
 
         buf[1] = 0x16;
-        assert_eq!(get_request_payload_len(&buf).unwrap(), Some(6));
+        assert_eq!(get_request_pdu_len(&buf).unwrap(), Some(7));
 
         buf[1] = 0x17;
         buf[10] = 99; // write byte count
-        assert_eq!(get_request_payload_len(&buf).unwrap(), Some(108));
+        assert_eq!(get_request_pdu_len(&buf).unwrap(), Some(109));
 
         buf[1] = 0x18;
-        assert_eq!(get_request_payload_len(&buf).unwrap(), Some(2));
+        assert_eq!(get_request_pdu_len(&buf).unwrap(), Some(3));
 
         // TODO: 0x2B
     }
 
     #[test]
-    fn test_get_response_payload_len() {
-
+    fn test_get_response_pdu_len() {
         let mut buf = BytesMut::new();
         buf.extend_from_slice(&[0x66, 0x01, 99]);
-        assert_eq!(get_response_payload_len(&buf).unwrap(),Some(100));
+        assert_eq!(get_response_pdu_len(&buf).unwrap(), Some(101));
 
         let mut buf = BytesMut::new();
         buf.extend_from_slice(&[0x66, 0x00, 99, 0x00]);
-        assert!(get_response_payload_len(&buf).is_err());
+        assert!(get_response_pdu_len(&buf).is_err());
 
         buf[1] = 0x01;
-        assert_eq!(get_response_payload_len(&buf).unwrap(), Some(100));
+        assert_eq!(get_response_pdu_len(&buf).unwrap(), Some(101));
 
         buf[1] = 0x02;
-        assert_eq!(get_response_payload_len(&buf).unwrap(), Some(100));
+        assert_eq!(get_response_pdu_len(&buf).unwrap(), Some(101));
 
         buf[1] = 0x03;
-        assert_eq!(get_response_payload_len(&buf).unwrap(), Some(100));
+        assert_eq!(get_response_pdu_len(&buf).unwrap(), Some(101));
 
         buf[1] = 0x04;
-        assert_eq!(get_response_payload_len(&buf).unwrap(), Some(100));
+        assert_eq!(get_response_pdu_len(&buf).unwrap(), Some(101));
 
         buf[1] = 0x05;
-        assert_eq!(get_response_payload_len(&buf).unwrap(), Some(4));
+        assert_eq!(get_response_pdu_len(&buf).unwrap(), Some(5));
 
         buf[1] = 0x06;
-        assert_eq!(get_response_payload_len(&buf).unwrap(), Some(4));
+        assert_eq!(get_response_pdu_len(&buf).unwrap(), Some(5));
 
         buf[1] = 0x07;
-        assert_eq!(get_response_payload_len(&buf).unwrap(), Some(1));
+        assert_eq!(get_response_pdu_len(&buf).unwrap(), Some(2));
 
         // TODO: 0x08
 
         buf[1] = 0x0B;
-        assert_eq!(get_response_payload_len(&buf).unwrap(), Some(4));
+        assert_eq!(get_response_pdu_len(&buf).unwrap(), Some(5));
 
         buf[1] = 0x0C;
-        assert_eq!(get_response_payload_len(&buf).unwrap(), Some(100));
+        assert_eq!(get_response_pdu_len(&buf).unwrap(), Some(101));
 
         buf[1] = 0x0F;
-        assert_eq!(get_response_payload_len(&buf).unwrap(), Some(4));
+        assert_eq!(get_response_pdu_len(&buf).unwrap(), Some(5));
 
         buf[1] = 0x10;
-        assert_eq!(get_response_payload_len(&buf).unwrap(), Some(4));
+        assert_eq!(get_response_pdu_len(&buf).unwrap(), Some(5));
 
         // TODO: 0x11
 
@@ -330,21 +377,21 @@ mod tests {
         // TODO: 0x15
 
         buf[1] = 0x16;
-        assert_eq!(get_response_payload_len(&buf).unwrap(), Some(6));
+        assert_eq!(get_response_pdu_len(&buf).unwrap(), Some(7));
 
         buf[1] = 0x17;
-        assert_eq!(get_response_payload_len(&buf).unwrap(), Some(100));
+        assert_eq!(get_response_pdu_len(&buf).unwrap(), Some(101));
 
         buf[1] = 0x18;
         buf[2] = 0x01; // byte count Hi
         buf[3] = 0x00; // byte count Lo
-        assert_eq!(get_response_payload_len(&buf).unwrap(), Some(258));
+        assert_eq!(get_response_pdu_len(&buf).unwrap(), Some(259));
 
         // TODO: 0x2B
 
         for i in 0x81..0xAB {
             buf[1] = i;
-            assert_eq!(get_response_payload_len(&buf).unwrap(), Some(1));
+            assert_eq!(get_response_pdu_len(&buf).unwrap(), Some(2));
         }
     }
 
@@ -354,7 +401,7 @@ mod tests {
 
         #[test]
         fn decode_partly_received_client_message() {
-            let mut codec = Codec::client();
+            let mut codec = ClientCodec::default();
             let mut buf = BytesMut::from(vec![
                 0x12, // server address
                 0x02, // function code
@@ -371,59 +418,101 @@ mod tests {
         }
 
         #[test]
-        fn decode_partly_received_server_message_0x16() {
-            let mut codec = Codec::server();
-            let mut buf = BytesMut::from(vec![
-                0x12, // server address
-                0x16, // function code
-                0x00, // irrelevant
-                0x00, // irrelevant
-            ]);
-            assert_eq!(buf.len(), MIN_ADU_LEN);
+        fn decode_empty_client_message() {
+            let mut codec = ClientCodec::default();
+            let mut buf = BytesMut::new();
+            assert_eq!(0, buf.len());
 
             let res = codec.decode(&mut buf).unwrap();
 
             assert!(res.is_none());
-            assert_eq!(buf.len(), MIN_ADU_LEN);
+            assert_eq!(0, buf.len());
+        }
+
+        #[test]
+        fn decode_single_byte_client_message() {
+            let mut codec = ClientCodec::default();
+            let mut buf = BytesMut::from(vec![0x00]);
+            assert_eq!(1, buf.len());
+
+            let res = codec.decode(&mut buf).unwrap();
+
+            assert!(res.is_none());
+            assert_eq!(1, buf.len());
+        }
+
+        #[test]
+        fn decode_empty_server_message() {
+            let mut codec = ServerCodec::default();
+            let mut buf = BytesMut::new();
+            assert_eq!(0, buf.len());
+
+            let res = codec.decode(&mut buf).unwrap();
+
+            assert!(res.is_none());
+            assert_eq!(0, buf.len());
+        }
+
+        #[test]
+        fn decode_single_byte_server_message() {
+            let mut codec = ServerCodec::default();
+            let mut buf = BytesMut::from(vec![0x00]);
+            assert_eq!(1, buf.len());
+
+            let res = codec.decode(&mut buf).unwrap();
+
+            assert!(res.is_none());
+            assert_eq!(1, buf.len());
+        }
+
+        #[test]
+        fn decode_partly_received_server_message_0x16() {
+            let mut codec = ServerCodec::default();
+            let mut buf = BytesMut::from(vec![
+                0x12, // server address
+                0x16, // function code
+            ]);
+            assert_eq!(buf.len(), 2);
+
+            let res = codec.decode(&mut buf).unwrap();
+
+            assert!(res.is_none());
+            assert_eq!(buf.len(), 2);
         }
 
         #[test]
         fn decode_partly_received_server_message_0x0f() {
-            let mut codec = Codec::server();
+            let mut codec = ServerCodec::default();
             let mut buf = BytesMut::from(vec![
                 0x12, // server address
                 0x0F, // function code
-                0x00, // irrelevant
-                0x00, // irrelevant
             ]);
-            assert_eq!(buf.len(), MIN_ADU_LEN);
+            assert_eq!(buf.len(), 2);
 
             let res = codec.decode(&mut buf).unwrap();
 
             assert!(res.is_none());
-            assert_eq!(buf.len(), MIN_ADU_LEN);
+            assert_eq!(buf.len(), 2);
         }
 
         #[test]
         fn decode_partly_received_server_message_0x10() {
-            let mut codec = Codec::server();
+            let mut codec = ServerCodec::default();
             let mut buf = BytesMut::from(vec![
                 0x12, // server address
                 0x10, // function code
-                0x00, // irrelevant
-                0x00, // irrelevant
             ]);
-            assert_eq!(buf.len(), MIN_ADU_LEN);
+            assert_eq!(buf.len(), 2);
 
             let res = codec.decode(&mut buf).unwrap();
 
             assert!(res.is_none());
-            assert_eq!(buf.len(), MIN_ADU_LEN);
+            assert_eq!(buf.len(), 2);
         }
 
         #[test]
         fn decode_rtu_message() {
-            let mut codec = Codec::client();
+            let mut codec = ClientCodec::default();
             let mut buf = BytesMut::from(vec![
                 0x01, // device address
                 0x03, // function code
@@ -436,24 +525,20 @@ mod tests {
                 0x9D, // crc
                 0x00,
             ]);
-            let RtuAdu { address, pdu } = codec.decode(&mut buf).unwrap().unwrap();
+            let ResponseAdu { hdr, pdu } = codec.decode(&mut buf).unwrap().unwrap();
             assert_eq!(buf.len(), 1);
-            assert_eq!(address, 0x01);
-            if let Pdu::Result(res) = pdu {
-                if let Response::ReadHoldingRegisters(data) = res.unwrap() {
-                    assert_eq!(data.len(), 2);
-                    assert_eq!(data, vec![0x8902, 0x42C7]);
-                } else {
-                    panic!("unexpected response")
-                }
+            assert_eq!(hdr.address, 0x01);
+            if let Ok(Response::ReadHoldingRegisters(data)) = pdu.into() {
+                assert_eq!(data.len(), 2);
+                assert_eq!(data, vec![0x8902, 0x42C7]);
             } else {
-                panic!("unexpected result")
+                panic!("unexpected response")
             }
         }
 
         #[test]
         fn decode_exception_message() {
-            let mut codec = Codec::client();
+            let mut codec = ClientCodec::default();
             let mut buf = BytesMut::from(vec![
                 0x66, //
                 0x82, // exception = 0x80 + 0x02
@@ -462,24 +547,24 @@ mod tests {
                 0x7E, // crc
             ]);
 
-            let RtuAdu { pdu, .. } = codec.decode(&mut buf).unwrap().unwrap();
-            if let Pdu::Result(res) = pdu {
-                let err = res.err().unwrap();
+            let ResponseAdu { pdu, .. } = codec.decode(&mut buf).unwrap().unwrap();
+            if let ResponsePdu(Err(err)) = pdu {
                 assert_eq!(format!("{}", err), "Modbus function 2: Illegal data value");
+                assert_eq!(buf.len(), 0);
             } else {
-                panic!("wrong pdu type");
+                panic!("unexpected response")
             }
-            assert_eq!(buf.len(), 0);
         }
 
         #[test]
         fn encode_read_request() {
-            let mut codec = Codec::client();
+            let mut codec = ClientCodec::default();
             let mut buf = BytesMut::new();
             let req = Request::ReadHoldingRegisters(0x082b, 2);
-            let pdu = Pdu::Request(req.clone());
+            let pdu = req.clone().into();
             let address = 0x01;
-            let adu = RtuAdu { address, pdu };
+            let hdr = Header { address };
+            let adu = RequestAdu { hdr, pdu };
             codec.encode(adu.clone(), &mut buf).unwrap();
 
             assert_eq!(
@@ -490,11 +575,12 @@ mod tests {
 
         #[test]
         fn encode_with_limited_buf_capacity() {
-            let mut codec = Codec::client();
+            let mut codec = ClientCodec::default();
             let req = Request::ReadHoldingRegisters(0x082b, 2);
-            let pdu = Pdu::Request(req.clone());
+            let pdu = req.clone().into();
             let address = 0x01;
-            let adu = RtuAdu { address, pdu };
+            let hdr = Header { address };
+            let adu = RequestAdu { hdr, pdu };
             let mut buf = BytesMut::with_capacity(40);
             unsafe {
                 buf.set_len(33);
