+++ src/frame/mod.rs
@@ -1,0 +1,162 @@
+#[cfg(feature = "rtu")]
+pub mod rtu;
+
+#[cfg(feature = "tcp")]
+pub mod tcp;
+
+use std::{error, fmt};
+
+/// A Modbus function code is represented by an unsigned 8 bit integer.
+pub(crate) type FunctionCode = u8;
+
+/// A Modbus address is represented by 16 bit (from `0` to `65535`).
+pub(crate) type Address = u16;
+
+/// A Coil represents a single bit.
+///
+/// - `true` is equivalent to `ON`, `1` and `0xFF00`.
+/// - `false` is equivalent to `OFF`, `0` and `0x0000`.
+pub(crate) type Coil = bool;
+
+/// Modbus uses 16 bit for its data items (big-endian representation).
+pub(crate) type Word = u16;
+
+/// Number of items to process (`0` - `65535`).
+pub(crate) type Quantity = u16;
+
+/// A request represents a message from the client (master) to the server (slave).
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum Request {
+    ReadCoils(Address, Quantity),
+    ReadDiscreteInputs(Address, Quantity),
+    WriteSingleCoil(Address, Coil),
+    WriteMultipleCoils(Address, Vec<Coil>),
+    ReadInputRegisters(Address, Quantity),
+    ReadHoldingRegisters(Address, Quantity),
+    WriteSingleRegister(Address, Word),
+    WriteMultipleRegisters(Address, Vec<Word>),
+    ReadWriteMultipleRegisters(Address, Quantity, Address, Vec<Word>),
+    Custom(FunctionCode, Vec<u8>),
+}
+
+/// The data of a successfull request.
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub enum Response {
+    ReadCoils(Vec<Coil>),
+    ReadDiscreteInputs(Vec<Coil>),
+    WriteSingleCoil(Address),
+    WriteMultipleCoils(Address, Quantity),
+    ReadInputRegisters(Vec<Word>),
+    ReadHoldingRegisters(Vec<Word>),
+    WriteSingleRegister(Address, Word),
+    WriteMultipleRegisters(Address, Quantity),
+    ReadWriteMultipleRegisters(Vec<Word>),
+    Custom(FunctionCode, Vec<u8>),
+}
+
+/// A server (slave) exception.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(crate) enum Exception {
+    IllegalFunction = 0x01,
+    IllegalDataAddress = 0x02,
+    IllegalDataValue = 0x03,
+    ServerDeviceFailure = 0x04,
+    Acknowledge = 0x05,
+    ServerDeviceBusy = 0x06,
+    MemoryParityError = 0x08,
+    GatewayPathUnavailable = 0x0A,
+    GatewayTargetDevice = 0x0B,
+}
+
+impl Exception {
+    pub(crate) fn description(&self) -> &str {
+        use crate::frame::Exception::*;
+
+        match *self {
+            IllegalFunction => "Illegal function",
+            IllegalDataAddress => "Illegal data address",
+            IllegalDataValue => "Illegal data value",
+            ServerDeviceFailure => "Server device failure",
+            Acknowledge => "Acknowledge",
+            ServerDeviceBusy => "Server device busy",
+            MemoryParityError => "Memory parity error",
+            GatewayPathUnavailable => "Gateway path unavailable",
+            GatewayTargetDevice => "Gateway target device failed to respond",
+        }
+    }
+}
+
+/// A server (slave) exception response.
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub(crate) struct ExceptionResponse {
+    pub(crate) function: FunctionCode,
+    pub(crate) exception: Exception,
+}
+
+/// Represents a message from the client (slave) to the server (master).
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub(crate) struct RequestPdu(pub(crate) Request);
+
+impl From<Request> for RequestPdu {
+    fn from(from: Request) -> Self {
+        RequestPdu(from)
+    }
+}
+
+impl From<RequestPdu> for Request {
+    fn from(from: RequestPdu) -> Self {
+        from.0
+    }
+}
+
+/// Represents a message from the server (slave) to the client (master).
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub(crate) struct ResponsePdu(pub(crate) Result<Response, ExceptionResponse>);
+
+impl From<Response> for ResponsePdu {
+    fn from(from: Response) -> Self {
+        ResponsePdu(Ok(from))
+    }
+}
+
+impl From<ExceptionResponse> for ResponsePdu {
+    fn from(from: ExceptionResponse) -> Self {
+        ResponsePdu(Err(from))
+    }
+}
+
+impl From<Result<Response, ExceptionResponse>> for ResponsePdu {
+    fn from(from: Result<Response, ExceptionResponse>) -> Self {
+        ResponsePdu(from.map(Into::into).map_err(Into::into))
+    }
+}
+
+impl From<ResponsePdu> for Result<Response, ExceptionResponse> {
+    fn from(from: ResponsePdu) -> Self {
+        from.0
+    }
+}
+
+impl fmt::Display for Exception {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "{}", self.description())
+    }
+}
+
+impl error::Error for Exception {
+    fn description(&self) -> &str {
+        self.description()
+    }
+}
+
+impl fmt::Display for ExceptionResponse {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "Modbus function {}: {}", self.function, self.exception)
+    }
+}
+
+impl error::Error for ExceptionResponse {
+    fn description(&self) -> &str {
+        self.exception.description()
+    }
+}
