+++ src/service/mod.rs
@@ -1,17 +1,17 @@
 #[cfg(feature = "rtu")]
 pub mod rtu;
+
 #[cfg(feature = "tcp")]
 pub mod tcp;
 
-/// An synchronous function from Request to a Response
-#[cfg(feature = "sync")]
-pub trait SyncService {
-    /// Requests handled by the service.
-    type Request;
-    /// Responses given by the service.
-    type Response;
-    /// Errors produced by the service.
-    type Error;
-    /// Process the request and return the response synchronously.
-    fn call(&mut self, _: Self::Request) -> Result<Self::Response, Self::Error>;
-}
+/// The Modbus address for sending a broadcast message to
+/// all connected stations (RTU) or slave devices (TCP).
+/// TODO: Sending broadcast messages is not supported yet!
+#[allow(dead_code)]
+pub(crate) const BROADCAST_ADDRESS: u8 = 0x00;
+
+/// The minimum Modbus address of a connected station (RTU) or slave device (TCP).
+pub(crate) const MIN_ADDRESS: u8 = 1;
+
+/// The maximum Modbus address of a connected station (RTU) or slave device (TCP).
+pub(crate) const MAX_ADDRESS: u8 = 247;
